Refactor: Fix VideoCard imports
