--- src/components/VideoCard/index.js
@@ -1,1 +1,0 @@
-export { default } from "./VideoCard";
